handled conflicts of base branch
